Extracted content generators to separate file
