+++ talkgenerator/schema/content_generators.py
@@ -1,0 +1,293 @@
+from schema.content_generator_structures import create_tracery_generator, create_templated_text_generator, \
+    GoodReadsQuoteGenerator, SplitCaptionsGenerator, create_reddit_image_generator, ShitPostBotURLGenerator, \
+    generate_google_image_generator, UnsplashURLGenerator, JobPrefixApplier, CountryPrefixApplier
+from sources import inspirobot, giphy, shitpostbot, google_images, unsplash
+from util.generator_util import CombinedGenerator, NoneGenerator, SeededGenerator, BackupGenerator, \
+    ExternalImageListGenerator, FromListGenerator, InvalidImagesRemoverGenerator, UnseededGenerator, TupledGenerator, \
+    MappedGenerator, InspiredTupleGenerator
+
+# ===============================
+# =====  CONTENT GENERATORS =====
+# ===============================
+
+# === TEXT GENERATORS ===
+
+# TITLES
+talk_title_generator = create_tracery_generator("data/text-templates/talk_title.json")
+talk_subtitle_generator = create_tracery_generator(
+    "data/text-templates/talk_subtitle.json"
+)
+
+
+def talk_title_generator_if_not_generated(presentation_context):
+    if presentation_context["title"] is not None:
+        return presentation_context["title"]
+    return talk_title_generator(presentation_context)
+
+
+default_slide_title_generator = create_templated_text_generator(
+    "data/text-templates/default_slide_title.txt"
+)
+
+default_or_no_title_generator = CombinedGenerator(
+    (1, default_slide_title_generator), (1, NoneGenerator())
+)
+
+anticipation_title_generator = create_templated_text_generator(
+    "data/text-templates/anticipation_title.txt"
+)
+
+conclusion_title_generator = create_templated_text_generator(
+    "data/text-templates/conclusion_title.txt"
+)
+inspiration_title_generator = create_templated_text_generator(
+    "data/text-templates/inspiration.txt"
+)
+anecdote_title_generator = create_templated_text_generator(
+    "data/text-templates/anecdote_title.txt"
+)
+history_title_generator = create_templated_text_generator(
+    "data/text-templates/history.txt"
+)
+history_person_title_generator = create_templated_text_generator(
+    "data/text-templates/history_person.txt"
+)
+history_and_history_person_title_generator = CombinedGenerator(
+    (4, history_title_generator), (6, history_person_title_generator)
+)
+about_me_title_generator = create_templated_text_generator(
+    "data/text-templates/about_me_title.txt"
+)
+
+# NAMES
+historical_name_generator = create_tracery_generator(
+    "data/text-templates/name.json", "title_name"
+)
+full_name_generator = create_tracery_generator(
+    "data/text-templates/name.json", "full_name"
+)
+
+# ABOUT ME
+_about_me_facts_grammar = "data/text-templates/about_me_facts.json"
+book_description_generator = create_tracery_generator(
+    _about_me_facts_grammar, "book_description"
+)
+location_description_generator = create_tracery_generator(
+    _about_me_facts_grammar, "location_description"
+)
+hobby_description_generator = create_tracery_generator(
+    _about_me_facts_grammar, "hobby_description"
+)
+job_generator = create_tracery_generator(_about_me_facts_grammar, "job")
+country_generator = create_tracery_generator(_about_me_facts_grammar, "country")
+
+# PROMPTS & CHALLENGES
+
+anecdote_prompt_generator = create_templated_text_generator(
+    "data/text-templates/anecdote_prompt.txt"
+)
+
+# QUOTES
+goodreads_quote_generator = GoodReadsQuoteGenerator(250)
+
+# DOUBLE CAPTIONS
+
+_double_image_captions_generator = create_templated_text_generator(
+    "data/text-templates/double_captions.txt"
+)
+_triple_image_captions_generator = create_templated_text_generator(
+    "data/text-templates/triple_captions.txt"
+)
+_historic_double_captions_generator = create_templated_text_generator(
+    "data/text-templates/historic_double_captions.txt"
+)
+
+double_image_captions_generator = SplitCaptionsGenerator(
+    _double_image_captions_generator
+)
+triple_image_captions_generator = SplitCaptionsGenerator(
+    _triple_image_captions_generator
+)
+historic_double_captions_generator = SplitCaptionsGenerator(
+    _historic_double_captions_generator
+)
+
+# Conclusions
+_conclusions_tuple_grammar = "data/text-templates/conclusion_tuple.json"
+conclusion_two_captions_tuple_generator = SplitCaptionsGenerator(
+    create_tracery_generator(_conclusions_tuple_grammar, "two_conclusions")
+)
+
+conclusion_three_captions_tuple_generator = SplitCaptionsGenerator(
+    create_tracery_generator(_conclusions_tuple_grammar, "three_conclusions")
+)
+
+# === IMAGE GENERATORS ===
+
+# INSPIROBOT
+inspirobot_image_generator = inspirobot.get_random_inspirobot_image
+
+# GIFS
+
+giphy_generator = SeededGenerator(
+    BackupGenerator(giphy.get_related_giphy, giphy.get_random_giphy)
+)
+reddit_gif_generator = create_reddit_image_generator(
+    "gifs", "gif", "gifextra", "nonononoYES"
+)
+
+combined_gif_generator = CombinedGenerator(
+    (1, giphy_generator), (1, reddit_gif_generator)
+)
+
+# REDDIT
+
+meme_reddit_image_generator = create_reddit_image_generator(
+    "meme",
+    "memes",
+    "MemeEconomy",
+    "wholesomememes",
+    "dankmemes",
+    "AdviceAnimals",
+    "comics",
+)
+weird_reddit_image_generator = create_reddit_image_generator(
+    "hmmm",
+    "hmm",
+    "wtf",
+    "wtfstockphotos",
+    "weirdstockphotos",
+    "darkstockphotos",
+    "photoshopbattles",
+    "confusing_perspective",
+    "cursedimages",
+    "HybridAnimals",
+    "EyeBleach",
+    "natureismetal",
+    "195",
+)
+
+neutral_reddit_image_generator = create_reddit_image_generator(
+    "Cinemagraphs",
+    "itookapicture",
+    "Art",
+    "artstore",
+    "pics",
+    "analog",
+    "ExposurePorn",
+    "Illustration",
+)
+
+shitpostbot_image_generator = ExternalImageListGenerator(
+    SeededGenerator(
+        BackupGenerator(
+            shitpostbot.search_images_rated, shitpostbot.get_random_images_rated
+        )
+    ),
+    ShitPostBotURLGenerator(),
+    weighted=True,
+)
+
+weird_punchline_static_image_generator = CombinedGenerator(
+    (4, weird_reddit_image_generator),
+    (6, shitpostbot_image_generator),
+    (1, meme_reddit_image_generator),
+)
+
+weird_punchline_image_generator = CombinedGenerator(
+    (10, weird_reddit_image_generator),
+    (8, shitpostbot_image_generator),
+    (6, combined_gif_generator),
+    (1, meme_reddit_image_generator),
+)
+
+# GOOGLE IMAGES
+
+generate_full_screen_google_image = generate_google_image_generator(
+    google_images.FullImageGenerator()
+)
+
+generate_wide_google_image = generate_google_image_generator(
+    google_images.WideImageGenerator()
+)
+
+generate_google_image = generate_google_image_generator(google_images.ImageGenerator())
+
+generate_google_image_from_word = FromListGenerator(
+    InvalidImagesRemoverGenerator(google_images.ImageGenerator())
+)
+
+# UNSPLASH
+
+generate_unsplash_image = ExternalImageListGenerator(
+    SeededGenerator(unsplash.search_photos),
+    UnsplashURLGenerator(),
+    check_image_validness=False,
+)
+generate_unsplash_image_from_word = ExternalImageListGenerator(
+    unsplash.search_photos, UnsplashURLGenerator(), check_image_validness=False,
+)
+
+neutral_image_generator = CombinedGenerator(
+    (1000, generate_unsplash_image),
+    (1, generate_google_image),
+    (300, neutral_reddit_image_generator),
+)
+
+neutral_image_generator_from_word = CombinedGenerator(
+    (1000, generate_unsplash_image_from_word),
+    (1, generate_google_image_from_word),
+    (300, UnseededGenerator(neutral_reddit_image_generator)),
+)
+
+neutral_or_weird_image_generator = CombinedGenerator(
+    (1, neutral_image_generator), (1, weird_punchline_image_generator)
+)
+
+# OLD/VINTAGE
+vintage_person_generator = create_reddit_image_generator("OldSchoolCool")
+vintage_picture_generator = create_reddit_image_generator(
+    "TheWayWeWere", "100yearsago", "ColorizedHistory"
+)
+
+reddit_book_cover_generator = create_reddit_image_generator(
+    "BookCovers", "fakebookcovers", "coverdesign", "bookdesign"
+)
+
+reddit_location_image_generator = create_reddit_image_generator(
+    "evilbuildings", "itookapicture", "SkyPorn", "EarthPorn"
+)
+
+# TUPLED ABOUT ME
+
+about_me_hobby_tuple_generator = TupledGenerator(
+    hobby_description_generator, weird_punchline_image_generator
+)
+about_me_book_tuple_generator = TupledGenerator(
+    book_description_generator, reddit_book_cover_generator
+)
+about_me_location_tuple_generator = TupledGenerator(
+    location_description_generator, reddit_location_image_generator
+)
+
+about_me_job_tuple_generator = MappedGenerator(
+    InspiredTupleGenerator(
+        MappedGenerator(job_generator, str.title), neutral_image_generator_from_word
+    ),
+    JobPrefixApplier(),
+)
+
+about_me_country_tuple_generator = MappedGenerator(
+    InspiredTupleGenerator(country_generator, neutral_image_generator_from_word),
+    CountryPrefixApplier(),
+)
+
+about_me_location_or_country_tuple_generator = CombinedGenerator(
+    (3, about_me_country_tuple_generator), (1, about_me_location_tuple_generator)
+)
+
+# Charts
+
+reddit_chart_generator = create_reddit_image_generator(
+    "dataisbeautiful", "funnycharts", "charts"
+)
