+++ talkgenerator/slide/slide_generator_types.py
